HOTFIX:fire:: review 수정 기능 오류 픽스

+++ src/main/java/com/iraefolio/service/ReviewService.java
@@ -42,14 +42,14 @@ public class ReviewService {
     /* UPDATE */
     public void update(List<ReviewEntity> entity) throws Exception {
 
+        log.error(entity);
+
         for (int i = 0; i < entity.size(); i++) {
             ReviewEntity e = entity.get(i);
-            if (e.isUpdated()) {
-                mapper.update(e);
-            }
-            if (e.isDeleted()) {
-                mapper.delete(e);
-            }
+
+            if (e.isUpdated()) mapper.update(e);
+
+            if (e.isDeleted()) mapper.delete(e);
         }
     }
 }
